Configure Rubocop
Adds the .rubocop.yml so that we can exclude automatically generated files from the inspected files. We don’t want these included because we want to be able to see the generated comments in case they need to be references later.

Updates several files to ensure they adhere to rubocops standards.