Update chart controller and associated tests
Removes the hard coded data from the previous feature where we setup the chart. This sets up the ability to contact the service that will be used to contact the API and get the up to date exchange rate data. Also, updates the tests to more accurately depict what the controller is responsible for.